프로젝트에 체크리스트 할당 구현
