+++ src/main/java/org/etmetmy/bn_server/domain/comment/service/CommentService.java
@@ -9,8 +9,8 @@ import org.etmetmy.bn_server.domain.comment.entity.Comment;
 import org.etmetmy.bn_server.domain.comment.repository.CommentRepository;
 import org.etmetmy.bn_server.domain.post.entity.Post;
 import org.etmetmy.bn_server.domain.post.repository.PostRepository;
-import org.etmetmy.bn_server.domain.user.entity.User;
-import org.etmetmy.bn_server.domain.user.repository.UserRepository;
+import org.etmetmy.bn_server.domain.entity.User;
+import org.etmetmy.bn_server.domain.repository.UserRepository;
 import org.etmetmy.bn_server.global.CustomException;
 import org.etmetmy.bn_server.global.StatusCode;
 import org.springframework.stereotype.Service;
