+++ src/main/java/org/etmetmy/bn_server/domain/dto/MemberUpdateRequest.java
@@ -1,0 +1,16 @@
+package org.etmetmy.bn_server.domain.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.etmetmy.bn_server.domain.entity.Role;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class MemberUpdateRequest {
+    private String name;
+    private String email;
+    private Long companyId;
+    private Role role;    // "ADMIN" or "USER"
+}
