+++ src/main/java/org/etmetmy/bn_server/domain/dto/request/UserLoginDto.java
@@ -1,0 +1,17 @@
+package org.etmetmy.bn_server.domain.dto.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class UserLoginDto {
+
+    private String email;
+    private String password;
+
+}
