+++ src/main/java/org/etmetmy/bn_server/domain/entity/ProjectMember.java
@@ -1,4 +1,4 @@
-package org.etmetmy.bn_server.domain.memo.entity;
+package org.etmetmy.bn_server.domain.entity;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
@@ -6,34 +6,29 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
 import org.etmetmy.bn_server.domain.project.entity.Project;
-import org.etmetmy.bn_server.domain.entity.User;
 import org.etmetmy.bn_server.global.entity.BaseEntity;
 
 @Entity
-@Table(name = "memo")
+@Table(name = "projectmember")
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
 @SuperBuilder
-public class Memo extends BaseEntity {
+public class ProjectMember extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "memo_id")
-    private Long memoId;
+    @Column(name = "project_member_id")
+    private Long projectMemberId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "project_id")
     private Project project;
 
-    @Column(name = "content", length = 1000)
-    private String content;
-
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
+    @JoinColumn(name = "user_id", nullable = false)
     private User user;
 
-    public void updateContent(String content) {
-        this.content = content;
-    }
+    @Column(name = "assigned_by")
+    private Long assignedBy;
 }
