+++ src/main/java/org/etmetmy/bn_server/domain/history/entity/HistoryComment.java
@@ -6,7 +6,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
 import org.etmetmy.bn_server.domain.comment.entity.Comment;
-import org.etmetmy.bn_server.domain.user.entity.User;
+import org.etmetmy.bn_server.domain.entity.User;
 import org.etmetmy.bn_server.global.entity.BaseEntity;
 
 @Entity
