+++ src/main/java/org/etmetmy/bn_server/domain/project/dto/entityDto/ProjectDTO.java
@@ -1,0 +1,9 @@
+package org.etmetmy.bn_server.domain.project.dto.entityDto;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class ProjectDTO {
+}
