+++ src/main/java/org/etmetmy/bn_server/domain/project/dto/request/ProjectAddCheckListRequest.java
@@ -1,0 +1,28 @@
+package org.etmetmy.bn_server.domain.project.dto.request;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.etmetmy.bn_server.domain.checkList.entity.CheckList;
+import org.etmetmy.bn_server.domain.project.entity.Project;
+import org.etmetmy.bn_server.domain.project.entity.ProjectCheckList;
+
+import java.util.List;
+
+@Getter
+@NoArgsConstructor
+public class ProjectAddCheckListRequest {
+
+    private List<Long>  checklistIds;
+
+    public static class Converter{
+
+        public static ProjectCheckList toEntity(Project project, CheckList checklist) {
+            return ProjectCheckList.builder()
+                    .project(project)
+                    .checkList(checklist)
+                    .answererId(null)
+                    .checked(false)
+                    .build();
+        }
+    }
+}
