+++ src/main/java/org/etmetmy/bn_server/domain/project/dto/request/ProjectCreateRequest.java
@@ -1,0 +1,13 @@
+package org.etmetmy.bn_server.domain.project.dto.request;
+
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.etmetmy.bn_server.domain.project.entity.Project;
+
+@Getter
+@NoArgsConstructor
+public class ProjectCreateRequest {
+
+}
