+++ src/main/java/org/etmetmy/bn_server/domain/project/entity/ProjectCheckList.java
@@ -1,0 +1,44 @@
+package org.etmetmy.bn_server.domain.project.entity;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.experimental.SuperBuilder;
+import org.etmetmy.bn_server.domain.checkList.entity.CheckList;
+import org.etmetmy.bn_server.domain.entity.User;
+import org.etmetmy.bn_server.global.entity.BaseEntity;
+
+@Entity
+@Table(name = "projectchecklist")
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@SuperBuilder
+public class ProjectCheckList extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "project_check_list_id")
+    private Long projectCheckListId;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "check_list_id", nullable = false)
+    private CheckList checkList;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "project_id")
+    private Project project;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "answerer_id")
+    private User answererId;
+
+    @Column(name = "checked")
+    private Boolean checked;
+
+    @PrePersist
+    protected void onCreate() {
+        if (checked == null) checked = false;
+    }
+}
