+++ src/main/java/org/etmetmy/bn_server/domain/project/repository/ProjectCheckListRepository.java
@@ -1,0 +1,9 @@
+package org.etmetmy.bn_server.domain.project.repository;
+
+import org.etmetmy.bn_server.domain.project.entity.ProjectCheckList;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ProjectCheckListRepository extends JpaRepository<ProjectCheckList, Long> {
+}
