+++ src/main/java/org/etmetmy/bn_server/domain/project/repository/ProjectMemberRespository.java
@@ -1,0 +1,8 @@
+package org.etmetmy.bn_server.domain.project.repository;
+
+
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ProjectMemberRespository {
+}
