+++ src/main/java/org/etmetmy/bn_server/domain/project/repository/ProjectRepository.java
@@ -1,0 +1,9 @@
+package org.etmetmy.bn_server.domain.project.repository;
+
+import org.etmetmy.bn_server.domain.project.entity.Project;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ProjectRepository extends JpaRepository<Project, Long> {
+}
