+++ src/main/java/org/etmetmy/bn_server/domain/project/repository/ProjectStageRepository.java
@@ -1,0 +1,7 @@
+package org.etmetmy.bn_server.domain.project.repository;
+
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ProjectStageRepository {
+}
