--- src/main/java/org/etmetmy/bn_server/domain/user/dto/MemberUpdateRequest.java
@@ -1,16 +1,0 @@
-package org.etmetmy.bn_server.domain.user.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.etmetmy.bn_server.domain.user.entity.Role;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class MemberUpdateRequest {
-    private String name;
-    private String email;
-    private Long companyId;
-    private Role role;    // "ADMIN" or "USER"
-}
