--- src/main/java/org/etmetmy/bn_server/domain/user/entity/ProjectMember.java
@@ -1,39 +1,0 @@
-package org.etmetmy.bn_server.domain.user.entity;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.experimental.SuperBuilder;
-import org.etmetmy.bn_server.domain.project.entity.Project;
-import org.etmetmy.bn_server.global.entity.BaseEntity;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Table(name = "projectmember")
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@SuperBuilder
-public class ProjectMember extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "project_member_id")
-    private Long projectMemberId;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "project_id")
-    private Project project;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id", nullable = false)
-    private User user;
-
-    @Column(name = "role", nullable = false, length = 255)
-    private String role;
-
-    @Column(name = "assigned_by")
-    private Long assignedBy;
-}
