--- src/main/java/org/etmetmy/bn_server/domain/user/service/CompanyService.java
@@ -1,11 +1,0 @@
-package org.etmetmy.bn_server.domain.user.service;
-
-
-import org.etmetmy.bn_server.domain.company.entity.Company;
-
-import java.util.Optional;
-
-public interface CompanyService {
-
-    Optional<Company> findByCompanyName(String companyName);
-}
