--- src/main/java/org/etmetmy/bn_server/domain/user/service/CompanyServiceImpl.java
@@ -1,19 +1,0 @@
-package org.etmetmy.bn_server.domain.user.service;
-
-import lombok.RequiredArgsConstructor;
-import org.etmetmy.bn_server.domain.company.entity.Company;
-import org.etmetmy.bn_server.domain.company.repository.CompanyRepository;
-import org.springframework.stereotype.Service;
-
-import java.util.Optional;
-
-@Service
-@RequiredArgsConstructor
-public class CompanyServiceImpl implements CompanyService{
-    private final CompanyRepository companyRepository;
-    @Override
-    public Optional<Company> findByCompanyName(String companyName) {
-        return companyRepository.findByCompanyName(companyName);
-
-    }
-}
