--- src/main/java/org/etmetmy/bn_server/domain/user/service/UserServiceImpl.java
@@ -1,106 +1,0 @@
-package org.etmetmy.bn_server.domain.user.service;
-
-import lombok.RequiredArgsConstructor;
-import org.etmetmy.bn_server.domain.company.entity.Company;
-import org.etmetmy.bn_server.domain.company.entity.CompanyType;
-import org.etmetmy.bn_server.domain.company.repository.CompanyRepository;
-import org.etmetmy.bn_server.domain.user.dto.MemberUpdateRequest;
-import org.etmetmy.bn_server.domain.user.dto.entity.UserDto;
-import org.etmetmy.bn_server.domain.user.dto.request.UserLoginDto;
-import org.etmetmy.bn_server.domain.user.entity.User;
-import org.etmetmy.bn_server.domain.user.repository.UserRepository;
-import org.etmetmy.bn_server.global.CustomException;
-import org.etmetmy.bn_server.global.StatusCode;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-
-@Service
-@RequiredArgsConstructor
-public class UserServiceImpl implements UserService{
-
-    private final UserRepository userRepository;
-    private final CompanyRepository companyRepository;
-    private final CompanyService companyService;
-    private final PasswordEncoder passwordEncoder;
-
-    // 회원 정보 수정 (Update)
-    @Transactional
-    public Long updateMember(Long memberId, MemberUpdateRequest request) {
-        // 1. 회원 찾기
-        User user = userRepository.findById(memberId)
-                .orElseThrow(() -> new CustomException(StatusCode.USER_NOT_FOUND));
-
-        // 2. 바꿀 회사 찾기
-        Company company = companyRepository.findById(request.getCompanyId())
-                .orElseThrow(() -> new CustomException(StatusCode.COMPANY_NOT_FOUND));
-
-        // 3. 정보 변경 (dirty checking)
-        user.updateInfo(request.getName(), request.getEmail(), company, request.getRole());
-
-        return user.getId();
-    }
-
-    // 2. 회원 삭제 (Delete)
-    @Transactional
-    public Long deleteMember(Long memberId) {
-        // 존재 여부 확인 후 삭제
-        User user = userRepository.findById(memberId)
-                .orElseThrow(() -> new CustomException(StatusCode.USER_NOT_FOUND));
-
-        Long userId = user.getId();
-        userRepository.delete(user);
-        return userId;
-    }
-
-
-    // 검색 & 조회 기능
-    public List<User> searchMembers(String name, String email, String companyName, String type) {
-        CompanyType companyType = null;
-
-        // 1. String 타입의 type 파라미터를 Enum으로 변환
-        if (type != null && !type.isBlank()) {
-            try {
-                companyType = CompanyType.valueOf(type.toUpperCase());
-            } catch (CustomException e) {
-                // 잘못된 type 값은 무시하고 null로 (쿼리에서 IS NULL 처리됨)
-            }
-        }
-
-        // 2. Repository의 @Query 메서드 호출
-        return userRepository.findByNamicMembers(
-                name,
-                email,
-                companyName,
-                companyType // Enum 값 전달
-        );
-    }
-
-    // UserService
-    public User login(UserLoginDto loginDto) {
-        User user = userRepository.findByEmail(loginDto.getEmail())
-                .orElseThrow(() -> new CustomException(StatusCode.USER_NOT_FOUND));
-
-        if (!passwordEncoder.matches(loginDto.getPassword(), user.getPassword())) {
-            throw new CustomException(StatusCode.PASSWORD_NOT_MATCH);
-        }
-
-        return user;
-    }
-    @Override
-    public Long joinUser(UserDto userDto) {
-        Company company = companyService.findByCompanyName(userDto.getCompany())
-                .orElseThrow(() -> new CustomException(StatusCode.COMPANY_NOT_FOUND));
-
-        User newUser = UserDto.Converter.toUser(userDto,company);
-
-        // 비밀번호 암호화
-        newUser.setPassword(passwordEncoder.encode(userDto.getPassword()));
-
-        Long userId = userRepository.save(newUser).getId();
-
-        return userId;
-    }
-}
