+++ src/main/java/org/etmetmy/bn_server/global/util/SessionUtil.java
@@ -1,8 +1,7 @@
 package org.etmetmy.bn_server.global.util;
 
 import jakarta.servlet.http.HttpSession;
-import org.etmetmy.bn_server.domain.user.entity.Role;
-import org.etmetmy.bn_server.domain.user.entity.User;
+import org.etmetmy.bn_server.domain.entity.User;
 import org.etmetmy.bn_server.global.CustomException;
 import org.etmetmy.bn_server.global.StatusCode;
 import org.etmetmy.bn_server.web.SessionConst;
