+++ src/test/java/org/etmetmy/bn_server/domain/user/entity/LoginTest.java
@@ -3,9 +3,11 @@ package org.etmetmy.bn_server.domain.user.entity;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.etmetmy.bn_server.domain.company.entity.Company;
 import org.etmetmy.bn_server.domain.company.repository.CompanyRepository;
-import org.etmetmy.bn_server.domain.user.dto.entity.UserDto;
-import org.etmetmy.bn_server.domain.user.dto.request.UserLoginDto;
-import org.etmetmy.bn_server.domain.user.service.UserService;
+import org.etmetmy.bn_server.domain.dto.entity.UserDto;
+import org.etmetmy.bn_server.domain.dto.request.UserLoginDto;
+import org.etmetmy.bn_server.domain.entity.Role;
+import org.etmetmy.bn_server.domain.entity.User;
+import org.etmetmy.bn_server.domain.service.UserService;
 import org.etmetmy.bn_server.web.SessionConst;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
